Move sunpy.cm to sunpy.visualization.colormaps
