--- docs/code_ref/cm.rst
@@ -1,16 +1,0 @@
-SunPy cm
-========
-
-This submodule implements solar physics specific colormaps. Upon import these
-colormaps are registered with matplotlib, so they can be used as if they were
-built in.
-
-A preview of all the SunPy colormaps is shown below:
-
-.. plot::
-
-   import sunpy.cm
-
-   sunpy.cm.show_colormaps()
-
-.. automodapi:: sunpy.cm
