+++ sunpy/visualization/colormaps/color_tables.py
